refactor: Add gvs package
- Improve code layout
