+++ .gitignore
@@ -1,1 +1,1 @@
-gvs
+bin/gvs
